Web Services REST

+++ AHMED-4TWIN5/src/main/java/tn/esprit/ahmedbenhmida4twin5/repositories/IInstructorRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.ahmedbenhmida4twin5.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.ahmedbenhmida4twin5.entities.Instructor;
 
-public interface IInstructorRepository extends JpaRepository<Instructor, Integer> {
+public interface IInstructorRepository extends JpaRepository<Instructor, Long> {
 }
