+++ AHMED-4TWIN5/src/main/java/tn/esprit/ahmedbenhmida4twin5/repositories/IPisteRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.ahmedbenhmida4twin5.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.ahmedbenhmida4twin5.entities.Piste;
 
-public interface IPisteRepository extends JpaRepository<Piste, Integer> {
+public interface IPisteRepository extends JpaRepository<Piste, Long> {
 }
