+++ AHMED-4TWIN5/src/main/java/tn/esprit/ahmedbenhmida4twin5/repositories/IRegistrationRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.ahmedbenhmida4twin5.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.ahmedbenhmida4twin5.entities.Registration;
 
-public interface IRegistrationRepository extends JpaRepository<Registration, Integer> {
+public interface IRegistrationRepository extends JpaRepository<Registration, Long> {
 }
