+++ AHMED-4TWIN5/src/main/java/tn/esprit/ahmedbenhmida4twin5/repositories/ISubscriptionRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.ahmedbenhmida4twin5.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.ahmedbenhmida4twin5.entities.Subscription;
 
-public interface ISubscriptionRepository extends JpaRepository<Subscription, Integer> {
+public interface ISubscriptionRepository extends JpaRepository<Subscription, Long> {
 }
